One pointer failure was dismissing the whole record, and it hid 112
report.py's bucket() returned one bucket per record, and ANY failure
matching PTR_WORDS sent the record to "pointer/memory (a missing
precondition, not a bug)". A translation unit reporting both

  dereference failure: pointer NULL in ...
  arithmetic overflow on signed shl in ...

was dismissed and its overflow was never printed to anybody. In the
kernel sweep that hid 9 signed +, 5 signed shl, 3 array upper bound, 2
signed -, 2 signed * and 2 array lower bound behind 176 NULL-pointer
preconditions - and an array bound is memory safety, not a missing
precondition; it is only in PTR_WORDS because CBMC spells its pointer
checks with "array " and "object".

The fix is the one the msun comment three lines below already describes
and which had never been applied to this case: STRIP the pointer
failures, then classify what is left. And print the failures that
SURVIVED the strip, because showing "dereference failure: pointer NULL"
under a heading that says arithmetic contradicts itself and may not
print the reason the record is there at all.

  worth a person's time   65 -> 177    (kernel 4 -> 23)

AND TWO OF THE 19 NEW KERNEL ONES ARE REAL

Both are arm64 header macros that shift into bit 31 of a signed int.
UBSan confirms both; one is a functional bug as well.

CPU_MATCH() CANNOT MATCH AN IMPLEMENTER >= 0x80

  #define CPU_IMPL_MASK  (0xff << 24)

is 4278190080, not representable in int - and as a signed int it is
NEGATIVE. CPU_MATCH compares

  ((mask) & PCPU_GET(midr)) == ((mask) & CPU_ID_RAW(...))

where the left side is int & uint64_t (mask sign-extends, midr's top
bits are RES0, result positive) and the right side is int & int (stays
negative, sign-extends at the ==). Measured, implementer 0xC0:

  lhs = 0x00000000c0000000
  rhs = 0xffffffffc0000000     CPU_MATCH -> 0

CPU_IMPL_AMPERE is 0xC0. It is only in identcpu.c's name table today,
which compares the extracted field directly, so no erratum is currently
misapplied - the next 0x80+ implementer needing a workaround would have
got one that silently never fired. With the U suffixes both sides read
0xc0000000; APM (0x50) is unchanged, checked.

THE SMC FUNCTION ID IS SIGN-EXTENDED INTO x0

SMCCC_FUNC_ID's `(type) << 31` has type = SMCCC_FAST_CALL = 1 for EVERY
SMCCC call in the tree. The negative int goes to psci_call(), whose
parameters are register_t - int64_t on arm64:

  old  fid as register_t = 0xffffffff80000000
  new  fid as register_t = 0x0000000080000000

Arm DEN 0028 puts the identifier in the low 32 bits as an unsigned
value. Monitors evidently ignore the high half, which is why nothing has
noticed.

Both registered in check_pbsd_marks.py and verified by reverting each
alone. The arm64 users still compile; sys/dev/psci/psci.c and
sys/arm64/arm64/machdep.c fail exactly as they did before the change,
checked by stashing it.

+++ hbsd/src/sys/dev/psci/smccc.h
@@ -36,11 +36,28 @@
 #define	SMCCC_VERSION_MAJOR(ver)	(((ver) >> 16) & 0x7fff)
 #define	SMCCC_VERSION_MINOR(ver)	((ver) & 0xffff)
 
+/*
+ * PBSD: unsigned. `(type) << 31` with type = SMCCC_FAST_CALL = 1 is
+ * `1 << 31` on a signed int - undefined behaviour (C11 6.5.7p4), and
+ * UBSan agrees: "left shift of 1 by 31 places cannot be represented in
+ * type 'int'". Every SMCCC function ID in the tree is a fast call, so
+ * every one of them goes through it.
+ *
+ * The value it happens to produce, 0x80000000 as a NEGATIVE int, is then
+ * passed to psci_call(), whose parameters are register_t - int64_t on
+ * arm64 - so it SIGN-EXTENDS to 0xffffffff80000000 and that is what goes
+ * into x0. Arm DEN 0028 makes the function identifier a 32-bit unsigned
+ * value in the low half of the register; monitors evidently ignore the
+ * high half, which is why nothing has ever noticed. Measured:
+ *
+ *   old  fid as register_t = 0xffffffff80000000
+ *   new  fid as register_t = 0x0000000080000000
+ */
 #define	SMCCC_FUNC_ID(type, call_conv, range, func)	\
-	(((type) << 31) |				\
-	 ((call_conv) << 30) |				\
-	 (((range) & 0x3f) << 24) |				\
-	 ((func) & 0xffff))
+	(((uint32_t)(type) << 31) |			\
+	 ((uint32_t)(call_conv) << 30) |		\
+	 (((uint32_t)(range) & 0x3f) << 24) |		\
+	 ((uint32_t)(func) & 0xffff))
 
 #define	SMCCC_YIELDING_CALL	0
 #define	SMCCC_FAST_CALL		1
